+++ src/mod.ts
@@ -1,2 +1,4 @@
-export * from './ip/mod.ts'
-export * from './socket.ts'
+export * from './ip.ts'
+export * from './v4/mod.ts'
+export * from './v6/mod.ts'
+export * from './socket/socket.ts'